Fix dynamicSource with container list
